pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.0.1"
description = "FastAPI Boilerplate Project"
authors = ["Mert Kutay <mertckutay@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
aerich = "^0.7.1"
aiohttp = { extras = ["speedups"], version = "^3.8.3" }
arrow = "^1.2.3"
babel = "^2.11.0"
boto3 = "^1.26.67"
emails = "^0.6"
fastapi = "^0.90.0"
geoip2 = "^4.6.0"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
orjson = "^3.8.6"
passlib = { extras = ["argon2"], version = "^1.7.4" }
pydantic = { extras = ["dotenv", "email"], version = "^1.10.4" }
pyhumps = "^3.8.0"
pytest-mock = "^3.10.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.5"
python-slugify = "^8.0.0"
redis = "^4.5.1"
secure = "^0.3.0"
sentry-sdk = "^1.15.0"
tenacity = "^8.2.0"
tortoise-orm = { extras = ["asyncpg"], version = "^0.19.3" }
user-agents = "^2.2.0"
uvicorn = { extras = ["standard"], version = "^0.20.0" }

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "^2.0.0"
black = "^23.1.0"
factory-boy = "^3.2.1"
faker = "^16.7.0"
httpx = "^0.23.3"
ipdb = "^0.13.11"
ipython = "^8.10.0"
isort = "^5.12.0"
mypy = "^1.0.0"
pre-commit = "^3.0.4"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-socket = "^0.6.0"
pytest-sugar = "^0.9.6"
ruff = "^0.0.244"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.core.config.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."

[tool.ruff]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "UP",  # pyupgrade
  "C",   # flake8-comprehensions
  "B",   # flake8-bugbear
  "ANN", # flake8-annotations
  "S",   # flake8-bandit
]
ignore = [
  "E501",   # line too long, handled by black
  "B008",   # do not perform function calls in argument defaults
  "ANN401", # *args, **kwargs Any type
  "ANN101", # self annotation
  "ANN102", # cls annotation
]

[tool.ruff.per-file-ignores]
"__init__.py" = [
  "F401", # imports in init files
]
"**/tests/*.py" = [
  "S101", # use of assert
]

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
files = ["app/**/*.py"]

[tool.pytest.ini_options]
addopts = ["--disable-socket", "--allow-unix-socket", "--allow-hosts=db"]

[tool.pyright]
autoSearchPaths = true
diagnosticMode = "openFilesOnly"
useLibraryCodeForTypes = true
typeCheckingMode = "basic"
reportPrivateUsage = false
reportPrivateImportUsage = false
